Add files via upload
This project is a command-line password manager implemented in Python. It provides users with a secure and convenient way to manage their passwords. The password manager uses AES encryption to store passwords, ensuring that they are protected from unauthorized access.

Key features of the password manager include:

Storing passwords: Users can add passwords for various websites or services, which are then encrypted and saved to a file.
Retrieving passwords: Users can retrieve their passwords by providing the master password used for encryption.
Generating passwords: The password manager can also generate strong, random passwords for users.
Master password security: The master password is used to derive an encryption key, which is used to encrypt and decrypt passwords. It is not stored anywhere, ensuring that only the user can access their passwords.
This project is designed for educational purposes and as a demonstration of basic password management principles. It may require further development and security enhancements for use in production environments.